lib/string/sprintf/: snprintf_(): Use ssize_t instead of size_t in $2
vsnprintf(3) returns an int.  By using ssize_t, which is also signed, we
avoid the need for a cast.

Cc: Serge Hallyn <serge@hallyn.com>

+++ lib/string/sprintf/snprintf.h
@@ -10,8 +10,8 @@
 
 #include <errno.h>
 #include <stdarg.h>
-#include <stddef.h>
 #include <stdio.h>
+#include <sys/types.h>
 
 #include "attr.h"
 #include "sizeof.h"
@@ -25,15 +25,14 @@
 
 
 format_attr(printf, 3, 4)
-inline int snprintf_(char *restrict s, size_t size, const char *restrict fmt,
-    ...);
+inline int snprintf_(char *restrict s, ssize_t size, const char *restrict fmt, ...);
 format_attr(printf, 3, 0)
-inline int vsnprintf_(char *restrict s, size_t size, const char *restrict fmt,
+inline int vsnprintf_(char *restrict s, ssize_t size, const char *restrict fmt,
     va_list ap);
 
 
 inline int
-snprintf_(char *restrict s, size_t size, const char *restrict fmt, ...)
+snprintf_(char *restrict s, ssize_t size, const char *restrict fmt, ...)
 {
 	int      len;
 	va_list  ap;
@@ -47,14 +46,14 @@ snprintf_(char *restrict s, size_t size, const char *restrict fmt, ...)
 
 
 inline int
-vsnprintf_(char *restrict s, size_t size, const char *restrict fmt, va_list ap)
+vsnprintf_(char *restrict s, ssize_t size, const char *restrict fmt, va_list ap)
 {
 	int  len;
 
 	len = vsnprintf(s, size, fmt, ap);
 	if (len == -1)
 		return -1;
-	if ((size_t) len >= size) {
+	if (len >= size) {
 		errno = E2BIG;
 		return -1;
 	}
